move DNS to a new mod
